Change `libssh2_session_init` to a function
Instead of a convenience macro, I have changed the
`libssh2_session_init` function to a "real" function. I am working on
a project with a very basic C FFI where only NULL is not adequately
supported and the `libssh2.h` header is ignored. To avoid issues with
NULL, I just changed the `libssh2_session_init` macro to a function.
I plan to do so for all of the `*_ex` functions.