Remove versioning from shared library
The version number suffix and symbolic linking are disabled for the
`labssh2` library. The `labssh2` shared library, labssh2.dll,
labssh2.dylib, or labssh2.so, is not intended to be "installed" or
distributed with a package manager for a system. It is intended to be
distributed with the LabSSH2 LabVIEW package and included in the install
location for LabVIEW, not the system. Thus, the version number suffix
and symbolic linking are not needed and only increase the "noise" of the
file system naming and distribution.